esmmlproject

Documentation for esmmlproject.

esm_ml_project.prepare_data_batches

esm_ml_project.prepare_pca_analysis

esm_ml_project.run_simulation

esm_ml_project.prepare_data_batches — Method

prepare_data_batches(t::AbstractVector, X::AbstractMatrix; train_frac=0.8, batchsize=10)

Splits data into train and validation sets and returns batches & separation index:

Returns

train_batches::Vector{Tuple{Vector, Matrix}}

valid_batches::Vector{Tuple{Vector, Matrix}}

split_idx::Int64`

esm_ml_project.prepare_pca_analysis — Method

prepare_pca_analysis(u_jet::Array{<:Real,3}, K::Int)

Perform PCA/EOF analysis on the input field u_jet (nlon, nlat, nt).

Arguments

u_jet : 3D array of the field with dimensions (nlon, nlat, nt).

K : Number of leading EOFs/PCs to retain.

Returns

A : (K × time) array of retained principal component time series.

PCs : (time × space) array of full PCs.

EOFs : (space × K) matrix of leading EOF patterns.

X_anom : (time × space) anomaly matrix.

X_timemean: (1 × space) temporal mean field at each gridpoint.

frac_var : fraction of variance explained by each EOF.

esm_ml_project.run_simulation — Function

run_simulation(spinup_days::Int; sim_days::Int=400, trunc::Int=31, nlayers::Int=8)

Run a Held–Suarez forced dry atmosphere simulation with SpeedyWeather.

Arguments

spinup_days: number of days to run without saving output (spinup).

sim_days: number of days to run with NetCDF output (default = 400).

trunc: spectral truncation T (default = 31).

nlayers: number of vertical layers (default = 8).

Returns

simulation: the final Simulation object from SpeedyWeather.